Add Table of Contents
Build a split panel layout with the table of contents in the left hand side and the documentation on the right. This uses flexbox for layout - browsers that do not support flexbox will fall back to a desktop version of the mobile layout.

IE11 does not currently display the flexbox layout despite supporting flexbox. This is because it does not support CSS feature detection (the @supports syntax).