Enable syntax highlighting
Use the middleman-syntax gem to automatically highlight blocks of code using [Rouge][1].

The HTML output that rouge creates is compatible with the python syntax highlighter ‘Pygments’. In order to ensure that the highlighting makes sense we build on top of [Base 16][2] to map pygments to a palette of sixteen colours.

The palette is then based loosely on both the GitHub colour palette and work that @edwardhorsford did, ensuring that all combinations meet colour contrast standards.

[1]: https://github.com/jneen/rouge
[2]: https://github.com/chriskempson/base16