Always show code block scrollbars (Webkit, Mac OS)
Unfortunately on Mac and iOS devices scrollbars are not by default visible until you scroll the container. This means it’s not always obvious when a container can be scrolled, as is the case with some code examples in the documentation.

There is a work around for this but it essentially involves re-styling the scrollbars. [1][2][3] This PR does just that, attempting to make them look as close to native Mac OS controls as possible.

[1]: http://stackoverflow.com/a/7855592
[2]: https://davidwalsh.name/osx-overflow
[3]: http://blog.0100.tv/2012/11/webkit-scrollbars-on-os-x/