Add "link" icon to headings
There's a desire for developers to be able to easily get a "deep link"
into a specific heading in the tech documentation. This implements a
GitHub-esque link icon which appears when hovering over a heading.

This has proven a bit of a faff – our headings are outputted by a
markdown parser (Redcarpet) which means we have to create a custom
Redcarpet renderer (which extends from the Middleman default, which
extends from Redcarpet::Renderer::HTML) and override the `header`
method.

This introduces further complexity, as Redcarpet builds its URL
fragments in a native extension (see
https://github.com/vmg/redcarpet/blob/master/ext/redcarpet/html.c#L274)
which can't be called from Ruby – only the `header` method is exposed.
Subsequently, there's a `header_anchor` method in the class which is
analogous to the Redcarpet implementation.

On the whole, I'm not 100% happy with this approach – it feels like we're
altering Redcarpet a bit too much for my liking.

The link icon in use is currently a Unicode link character (🔗) – this
will be updated once there's consensus on which icon we should actually
be using.