Add "link" icon to headings with JavaScript
See alphagov/tech-docs-template#11 for background.

This adds some JavaScript to inject an anchor into each heading tag on
page load. The rationale here is that ultimately if this doesn't work
due to the user having JS disabled or any other reason, it won't break
the core experience of the website – it's an enhancement.

The alternative, rendering these in Ruby, introduces a reasonable amount
of technical debt and also raises some questions around accessibility.