Add Jasmine for JS testing
We're keen to do some testing of the JavaScript bits and bobs in this
project, and subsequently need a test framework to do so. Jasmine
appears to be the weapon of choice in other parts of GDS (including the
service manual), so it makes sense to use it here, too.

This adds a very basic Jasmine configuration that we can write some
tests on top of.