Add a max-width to images
At the moment, an image inserted into the tech docs will be rendered at
whatever resolution it's supplied in. This is somewhat far from ideal,
as a massive image could end up breaking the page in unpleasant ways.

This adds a max-width of 100% to all images inside the technical
documentation container to ensure we stop anything overflowing it.