Automatically linkify images
It might be nice for us to automatically wrap images in a link, so if
they're shrunk for whatever reason they can be clicked on to show a
full-resolution version. This does just that.

There's not an option built in to Redcarpet which allows us to do this,
so this adds a custom Redcarpet renderer which wraps up images in an <a>
tag.