Add fixed table of contents header on mobile
As the technical documentation is a very long document, there is a need to be able to easily jump around it when on using a mobile device or a narrow viewport.

This adds a ‘sticky’ ‘Table of Contents’ element when within the mobile breakpoint which, when clicked, expands into a scrollable table of contents. It makes use of the very new `position: sticky` CSS attribute which is polyfilled for older browsers (which is, in fact, most of them).

This also adds an autoprefixer to the asset pipeline.