Fix navigation toggle stacking bug
The navigation toggle has a z-index of 10 but this is ignored when the fixed behaviour is being polyfilled, because it falls back to static positioning where z-index is ignored.

This meant that the negative margin on the headings (which is in place so that when using the navigation the heading appears within the visible part of the page and not behind the navigation toggle) obscured the navigation toggle and meant it did not work when at the top of the page.

By wrapping the toggle with another element and making _that_ element ‘sticky’, and giving the toggle element relative positioning and a z-index, we can avoid this behaviour.