Fix layout in IE8 through IE10
IE8 through IE10 don’t support flexbox, so fall back to a simpler layout with the navigation floated to the left.

We discussed moving away from flexbox to just using absolutely positioned divs to define the application layout, but it would require us to ‘know’ and fix the the size of the header. Given the header currently ‘breaks’ onto two lines for the Platform as a Service documentation, this won’t really work for us.

We are making an assumption that IE8-IE10 will account for a even lower percentage of users than we see on GOV.UK (~2.75%) as they will be from a more technical background. We should validate this once we have collected analytics data in production.