Add some basic configuration options
In an attempt to make the 'templatization' of this project a bit easier,
this adds a configuration hash in `config.rb` to set some options that
are likely to change between projects.

I'm not entirely convinced by the header links being in here – it does
nicely abstract them away from the actual HTML, but open to better
alternatives.