Fix duplicated use of ‘#toc’ id, move classes to nav
In an ideal world the `<ul>` can be simply generated by the table of contents helper, and this means we don’t want to e.g. add classes to it.

Those classes would be just as happy living on the parent `<nav>` element, and this also fixes the fact that two elements have the `#toc` id. Whoops.