Add rel="noopener" when linking to images
We're using target="_blank" to open images in a new window. There are known issues with being able to exploit such links using `window.opener' (https://mathiasbynens.github.io/rel-noopener/)

Whilst the images are *probably* going to be hosted within the same repo and under our control, we can't really guarantee it. Adding rel="noopener" is not detrimental, so let's play it safe.