Fix rubber banding in iOS.
By default when the table of contents is at the top or bottom, scrolling in that direction will scroll the body 'behind' the table of contents. Fix this by preventing ever reaching the top or bottom of the table of contents (by 1 pixel) using javascript [1]. Non-JS users will just experience the (slightly) funky scrolling behaviour.

[1]: http://blog.christoffer.me/six-things-i-learnt-about-ios-safaris-rubber-band-scrolling/