Allow middleman init to be run with options
The `middleman init` template script is used to create new sites and to
update existing sites.

This commit allows users to configure the `middleman init` template
script with environment variables. This makes it possible to create and
build a site programatically on CI.

It also allows the update script to be run like:

```
FIRST_TIME=false middleman init . -T alphagov/tech-docs-template
```