Add unique identifier generator
This generates unique identifiers for headings, allowing for the same heading to appear multiple times in the document.

At the minute the difficulty with this is that we need to understand the context. When running `middleman build` this works fine, but when using `middleman server` the headings are seen as duplicated when the page is reloaded, because there’s nothing to bind the state to other than the renderer which is not recreated between requests.