Remember in page scroll state
Because we are using a scrollable div rather than scrolling the document body itself, browser behaviour is not quite what you’d expect. For example:

1) When navigating to an external link, using the browser back button takes you back to the top of the documentation, rather than remembering where you were.
2) When navigating the document using the table of contents, clicking the back button has no effect at all.

This fixes both of those problems by continuously storing the scroll offset using the browser’s history API as the user scrolls. Then, on initial page load (refreshing, or navigating back to the document from an external link) and whenever a `popstate` event is fired (the user clicks back after clicking an internal link), the scroll offset from any previous history state is restored.

We use `requestAnimationFrame` to debounce scroll events, polyfilling this for older browsers.

In order to maintain consistency between the URL that’s displayed in the browser and the part of the page the user is returned to, we also keep the two in sync by re-writing the URL as part of that same `replaceState` call. The fact that we need to know this is part of the reason why this code is not a separate module as you might expect.