Fix scroll history bug
If the user navigates within the page before they scroll, the scroll position will not be recorded in their history state object so clicking the back button will not take them back to the top of the page as they might expect it to.

We can work around this by scrolling to the top of the page when an empty state is popped from the history api.