Store initial scroll state on page load
The previous solution worked in Chrome, but Firefox and Safari fire pop state events when a anchor link is clicked, which means you can’t use the navigation. At all. Oops.

Instead of assuming that when there is no scrollTop stored we should scroll to the top, explicitly store the scroll position on page load if no existing state is present. Also allow restoreScrollPosition to scroll to the top as a value of 0 for scrollTop would be falsey.