Fix initial page position when linking to anchor
The call to handleScrollEvent was previously causing the browser to jump back up to the start of the document. This seems to be because:

1) handleScrollEvent is called _before_ the browser has scrolled the user to the target heading
2) tocItemForFirstElementInView therefore returns the first heading on the page
3) the current history state is therefore updated with the hash of the first heading on the page
4) this appears to be enough to cause the browser to scroll back to the top of the document

We need slightly different behaviour when loading the page for the first time - if there is an anchor specified, we should use the hash and offset for that entry rather than trying to find the first element in view. We’ll fall back to the first element if there is no hash, or the hash does not link to somewhere in the ToC.