Fix in page navigation on mobile
On mobile, the $tocItems array included the close link, which has an href of `#`, which is an invalid selector to try and search for. This fixes that by only including links within `js-toc-list`, which excludes the close link.

Also on mobile, `tocItemForFirstElementInView` does not find anything. We then try and call attr(`href`) on null within `storeCurrentPositionInHistoryApi`. This guards against $activeTocItem being falsey.