Use javascript menu toggle on nav, not a checkbox
The menu show/hide checkbox ’hack’ allows us to show and hide the menu without Javascript enabled, but after further research there are some browser support and accessibility issues with this approach [1] which means a more ‘traditional’ javascript based approach makes more sense.

This changes the toggle into a button which adds a `nav-open` class to the HTML element, which we can then use to conditionally style the navigation to show or hide it. It also ensures that if Javascript is not enabled, or fails to load, the navigation menu is displayed and the toggle button is hidden.

[1]: http://dev.edenspiekermann.com/2016/03/07/introducing-a11y-toggle/