Bind aria attributes to actual behaviour of ToC
We need the aria attributes to be applied responsively - on desktop the navigation is always visible. The easiest way to do this is to actually check whether the ToC is visible or not.

We also need to update the aria attributes when the window is resized as we may have changed from mobile to desktop, or vice versa.