Remove text from anchor links
When Google decides to use the h1 of the document, it’s extracting ‘Link to here’ as part of the heading and including it in the title of the page in search results.

Remove ‘Link to here’ as it’s hidden from screen readers anyway, and users of assisted tech would be able to ‘create’ links to individual headings using the table of contents.