Guard against no heading being seen as active
On mobile, we sometimes (?) don’t find any of the possible targets to be at the top of the screen, because their offset is the top of the page and there is a header between the first target and that.

Guarding to allow for result being nil fixes this. It means we don’t update the hash or the nav, but in the mobile viewport that’s not a problem.