Revise highlight state for table of contents
Rather than emboldening the active element, we highlight it with a light grey background and add a five pixel left hand border in the active link colour. This makes the highlighting more obvious, and allows us to do away with the letter-spacing ‘hack’ which was trying to prevent items reflowing when they became bold.

Unfortunately because each item in the table of contents now needs to effectively act as a block of the same size, we can no longer rely on the nested nature of the lists. Instead we have to define the behaviour for each level individually. We’ve implemented indentation for the first six levels of navigation. In a real world situation this number of levels are unlikely to be exposed, but given we have to choose a number of levels to support, matching the h1…h6 elements makes some sense.