Correct configuration for GA
Set commands should be added to the command queue after the create tracker command.

To avoid missing hits in GA when a document unloads, set the tracker to use the Navigator.sendBeacon() transport mechanism when available.  This will transmit hits to GA asynchronously when the User Agent has an opportunity to do so, without delaying the unload or affecting the performance of the next navigation.  GA will fallback to XMLHttpRequest or Image transport mechanisms when navigator.sendBeacon isn't supported by the user's browser.

See:
https://developer.mozilla.org/en-US/docs/Web/API/Navigator/sendBeacon

https://developers.google.com/analytics/devguides/collection/analyticsjs/sending-hits#specifying_different_transport_mechanisms