Add support for middleman using Windows 64-bit 'mingw32' platform
We want users who have installed Ruby using RubyInstaller x64 (or a
64-bit mingw installation) to be able to use `middleman init -T
alphagov/tech-docs-template`; currently they will get warnings saying:

> The dependency ... will be unused by any of the platforms Bundler is
installing for. Bundler is installing for x64-mingw32 but the dependency
is only for x86-mswin32, x86-mingw32.

because we only specify `mingw` in the [template Gemfile](template/Gemfile).
Bundler treats `mingw` as 32-bit only [[1]]. Without `wdm` the template
Thorfile will fail to execute with the error

> Unable to find a spec satisfying wdm (~> 0.1.0) in the set.

This commit adds `x64_mingw` to the list of platforms which require
`wdm` and `tzinfo-data`, which should fix things.

[1]: https://bundler.io/man/gemfile.5.html#PLATFORMS