Rename default branch to 'main'
We've renamed the default branch to main, following best practice on
GitHub [[1]], and we want and expect that users will do the same. This
is especially the case for new repositories on GitHub, which will have
the name 'main' instead of 'master'.

This commit updates the example deploy script to match.

We also add the new default branch name to the config template. This is
because in the tech-docs-gem we assume the default branch is named
'master', and changing that would be a breaking change. By changing it
here we can immediately reach new users of the tech docs template
(without affecting our ability to make the breaking change), and it also
keeps the example deploy script and the build output consistent.

[1]: https://github.com/github/renaming